refactor: update dir layouts
