+++ .gitignore
@@ -88,7 +88,7 @@ docs
 flow-typed
 typings
 raw
-*.tsbuildinfo
+buildcache
 buildstamp.json
 
 # Autogenerated
