+++ packages/abstract/src/main/ts/aliases.ts
@@ -1,8 +1,4 @@
 /** @module @qiwi/substrate-abstract */
 /** */
 
-export {
-  AError as AbstractError,
-  IError as Error,
-  IErrorOpts as ErrorOpts,
-} from './AError'
+export { AError as AbstractError, IError as Error, IErrorOpts as ErrorOpts } from './AError'
