+++ packages/abstract/src/test/dts/index.d.ts
@@ -1,0 +1,1 @@
+export * from '../../../target/esm/index';
