--- packages/abstract/src/test/ts/libdef/index.ts
@@ -1,1 +1,0 @@
-export * from '@qiwi/substrate-abstract'
