+++ packages/infra/package.json
@@ -47,6 +47,7 @@
     "terser": "^5.16.3",
     "ts-jest": "^29.0.5",
     "ts-node": "10.9.1",
+    "tsc-esm-fix": "^2.20.12",
     "typedoc": "^0.23.24",
     "typescript": "4.9.x"
   },
