+++ packages/std/jest.config.json
@@ -9,7 +9,7 @@
     "<rootDir>/packages/std/src/main/ts/*.ts"
   ],
   "testMatch": [
-    "<rootDir>/packages/std/src/test/ts/unit/*.ts"
+    "<rootDir>/packages/std/src/test/ts/*.ts"
   ],
   "testPathIgnorePatterns": [],
   "transformIgnorePatterns": [
