+++ packages/std/src/test/ts/stack.ts
@@ -1,4 +1,4 @@
-import { Stack } from '../../../main/ts'
+import { Stack } from '../../main/ts'
 
 describe('stack', () => {
   describe('constructor', () => {
