+++ packages/types/jest.config.json
@@ -10,7 +10,7 @@
     "<rootDir>/packages/types/src/main/**/*.ts"
   ],
   "testMatch": [
-    "<rootDir>/packages/types/src/test/ts/unit/**/*.ts"
+    "<rootDir>/packages/types/src/test/ts/**/*.ts"
   ],
   "transformIgnorePatterns": [
     "/!node_modules\\/lodash-es/",
