+++ packages/types/package.json
@@ -3,21 +3,22 @@
   "version": "1.55.1",
   "exports": {
     ".": {
-      "import": "./target/cjs/index.js",
-      "require": "./target/esm/index.js",
+      "import": "./target/esm/index.js",
+      "require": "./target/cjs/index.js",
       "types": "./target/esm/index.d.ts"
     }
   },
+  "types": "./target/esm/index.d.ts",
   "scripts": {
     "clean": "rimraf target buildcache flow-typed typings docs coverage buildstamp.json",
-    "build": "yarn aliasgen && concurrently -n yarn:build:cjs yarn:build:esm yarn:build:ts yarn:docs yarn:build:stamp",
+    "build": "yarn aliasgen && concurrently -n yarn:build:cjs yarn:build:ts yarn:docs yarn:build:stamp && yarn build:esm",
     "build:cjs": "mkdirp target/cjs && tsc -p tsconfig.json --target ES5 --module commonjs --outDir target/cjs",
     "build:esm": "mkdirp target/esm && tsc -p tsconfig.json --target ES2020 --module es2020  --outDir target/esm",
     "build:ts": "mkdirp target/ts && cp -R ./src/main/ts/. ./target/ts/.",
     "build:stamp": "buildstamp --git --out.path buildstamp.json",
     "test": "yarn test:dts",
     "test:unit": "yarn jest --config=jest.config.json --detectOpenHandles --forceExit",
-    "test:dts": "yarn dtslint src/test/ts --localTs ../../node_modules/typescript/lib --expectOnly",
+    "test:dts": "yarn dtslint src/test/dts --localTs ../../node_modules/typescript/lib --expectOnly",
     "aliasgen": "node ../../scripts/js/alias-gen.js",
     "docs": "typedoc"
   },
