+++ packages/types/src/test/dts/ICollection.ts
