+++ packages/types/src/test/dts/IConfig.ts
