+++ packages/types/src/test/dts/IConfigurable.ts
