+++ packages/types/src/test/dts/ICurrency.ts
