+++ packages/types/src/test/dts/IDeviceInfo.ts
