+++ packages/types/src/test/dts/INormalValue.ts
