+++ packages/types/src/test/dts/IPromise.ts
