+++ packages/types/src/test/dts/IStorage.ts
