+++ packages/types/src/test/dts/IStringMap.ts
