+++ packages/types/src/test/dts/TPredicate.ts
